% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{read_obj}
\alias{read_obj}
\title{Read OBJ}
\usage{
read_obj(x, ...)
}
\arguments{
\item{x}{source of a OBJ model}

\item{...}{ignored}
}
\value{
an `obj` structure
}
\description{
Read Wavefront OBJ format
}
\details{
The `obj` structure contains the face and vertex and normal data from a .obj file,
most of it anyway.
}
\examples{
f <- system.file("extdata/greek_bust.obj.zip", package = "obj")
obj <- read_obj(f)
}
